FF-328 #comment moved auth0 mixin into app
